requirements.txt
torch
torchvision
rdkit
pubchempy
datasets
numpy
pandas
rembg
requests
pillow
tqdm
opencv-python
